Edited lots of stuff:
Nodes and segments will now be stored in lists.
Nodes are now aware of the segments attached to them.
Slightly changed the structure of the Node and PipeSegment classes.
Added sample file to pull element classes from.
Added method to PipeSegment class for calculating head loss.
Added generic methods to PipeSegment and Node for adding elements or
attributes.